chameleon/micro-benchmarks/requirements.txt
dtaidistance==2.3.10
fastdtw==0.3.4
filterpy==1.4.5
idna==3.3
kaleido==0.2.1
kiwisolver==1.4.2
linmdtw==0.1.7
plotly==5.3.1
psutil==5.9.4
pymjc==1.0.3
requests==2.28.2
seaborn==0.11.2
shortuuid==1.0.9
tikzplotlib==0.10.1
tomli==1.2.3
tornado==6.1
urllib3==1.26.9
ruptures==1.1.7
kats==0.2.0
wfdb==4.1.0
paho-mqtt==1.6.1